requirements.txt
scikit-learn
tensorflow
numpy
pandas
opencv-python
matplotlib
nengo
pytorch
spikingjelly
keras
